+++ src/main/java/frc/robot/constants/RobotMap.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot;
+package frc.robot.constants;
 
 import java.util.HashMap;
 import java.util.Hashtable;
@@ -14,8 +14,8 @@ import edu.wpi.first.math.geometry.Translation2d;
 import edu.wpi.first.math.kinematics.SwerveDriveKinematics;
 import edu.wpi.first.math.util.Units;
 
-/** Add your docs here. */
-public final class Constants {
+/** CAN ID, Invert, Pod Positions, Offsets, Conversion Rates */
+public final class RobotMap {
     public static final int leftAzimuthID = 2;
     public static final int rightAzimuthID = 3;
 
@@ -27,46 +27,24 @@ public final class Constants {
 
     public static final int pigeonID = 6;
 
-    public static final double leftAbsoluteEncoderOffset = 3.45;//absolute encoder reading at position
-    public static final double rightAbsoluteEncoderOffset = 5.420739537358221 - 4.705;// gears facing inwards: fwd/bck TODO: less janky alignment
-
     public static final boolean leftAzimuthInvert = false;
     public static final boolean rightAzimuthInvert = false;
     public static final boolean leftDriveInvert = false;
     public static final boolean rightDriveInvert = true;
-
-    public static final double azimuthRadiansPerRotation = 2*Math.PI*15/33;
-    public static final double driveMetersPerRotation = Units.inchesToMeters(2) * Math.PI * 33 / 45 / 2;
-    public static final double absoluteEncoderRadiansPerRotation = 2*Math.PI;
-    
-    public static final int azimuthAmpLimit = 35;
-    public static final int driveAmpLimit = 25;
-    public static final int driveTopAmpLimit = 90;
-
-    public static final double robotMaxSpeed = 0.25;
-    public static final double podMaxSpeed = 1;
-    public static final double azimuthMaxOutput = 0.25;
-    public static double driveSpeedToPower = 1.0;
-
-    public static final double azimuthkP = 0.4;
-    public static final double azimuthkI = 0.0;
-    public static final double azimuthkD = 0.003;
-    public static final double azimuthkIz = 0;
-    public static final double azimuthDriveSpeedMultiplier = 0;//0.5;
-
-    public static final IdleMode azimuthMode = IdleMode.kBrake;
-    public static final IdleMode driveMode = IdleMode.kCoast;
-
-    public static final double gyroP = 2;
-    public static final double gyroI = 0.0;
-    public static final double gyroD = 0.00;
 
     public static final Translation2d robotCenter = new Translation2d(0, 0); // serves as "center of robot for calculations; robot will turn about this point
     public static final Translation2d leftPodPosition = new Translation2d(-Units.inchesToMeters(5.5), -Units.inchesToMeters(5.5)); // units in meters
     public static final Translation2d rightPodPosition = new Translation2d(Units.inchesToMeters(5.5), Units.inchesToMeters(5.5));
     public static final SwerveDriveKinematics drivetrainKinematics = new SwerveDriveKinematics(robotCenter.minus(leftPodPosition), robotCenter.minus(rightPodPosition));
 
-    public static final int driverPort = 0;
-    public static final double driveMotorRampRate = 0.2;
+    public static final double leftAbsoluteEncoderOffset = 3.45;//absolute encoder reading at position
+    public static final double rightAbsoluteEncoderOffset = 5.420739537358221 - 4.705;// gears facing inwards: fwd/bck TODO: less janky alignment
+
+    public static final double azimuthRadiansPerMotorRotation = 2*Math.PI*15/33;
+    public static final double driveMetersPerMotorRotation = Units.inchesToMeters(2) * Math.PI * 33 / 45 / 2;
+    public static final double absoluteRadiansPerEncoderRotation = 2*Math.PI;
+    
+
+    
 
 }
